requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
bidict==0.23.1
boto3==1.34.0
botocore==1.34.162
cachetools==5.5.2
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.7
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.108.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
h11==0.14.0
httplib2==0.22.0
icalendar==5.0.11
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
passlib==1.7.4
pillow==10.2.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.1.17
psycopg-pool==3.2.4
pyasn1==0.5.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-multipart==0.0.6
python-socketio==5.11.0
requests==2.32.3
requests-oauthlib==2.0.0
resend==2.6.0
rsa==4.9
s3transfer==0.9.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3>=1.25.4,<1.27
uvicorn==0.25.0
wsproto==1.2.0
